Improves mod loading and UI
Refactors mod loading logic and introduces a new UI for managing and displaying mods.
This includes a prefab for mod list entries, a Lua manager for mod scripting, and utilities for mod data loading.

Removes the old Metronome system and associated files in favor of a more modular and extensible system.

Updates file structure to improve organization.